requirements.txt
asgiref==3.5.0
asn1crypto==0.24.0
backports.zoneinfo==0.2.1
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.10
crypto==1.4.1
cryptography==36.0.1
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-mathfilters==1.0.0
djangorestframework==3.12.2
enum34==1.1.6
gevent==21.1.2
greenlet==1.0.0
html5lib==1.1
idna==2.6
ipaddress==1.0.17
jwt==1.3.1
keyring==10.6.0
keyrings.alt==3.0
Naked==0.1.31
olefile==0.45.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.13.0
pycryptodomex==3.13.0
PyJWT==1.7.1
PyOpenGL==3.1.0
Pyphen==0.10.0
pytz==2021.1
pyxdg==0.25
PyYAML==5.4.1
requests==2.25.1
SecretStorage==2.3.1
shellescape==3.8.1
six==1.11.0
sqlparse==0.4.1
stripe==2.56.0
tinycss2==1.1.0
urllib3==1.26.3
WeasyPrint==52.2
webencodings==0.5.1
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.2.0
